requirements.txt
# # Core
# pydantic>=2.6
# pydantic-settings>=2.2
# SQLAlchemy>=2.0
# psycopg[binary]>=3.2

# # Scraping
# httpx>=0.27
# beautifulsoup4>=4.12
# lxml>=5.2
# tenacity>=8.1.0,<9

# # Testing
# pytest>=8.2

# # Logging
# pyyaml>=6.0

# Core / DB / ETL
pandas>=2.0
SQLAlchemy>=2.0
psycopg[binary]>=3.1
psycopg2-binary>=2.9
python-dotenv>=1.0
tqdm>=4.0

# API
fastapi>=0.95
uvicorn[standard]>=0.22

# Scraping (if used)
requests>=2.30
beautifulsoup4>=4.12
lxml>=4.9
httpx>=0.27
tenacity>=8.1.0,<9

# Utilities & testing
pytest>=8.2
pydantic>=2.2
pyyaml>=6.0
joblib>=1.3  # if you persist/load models
numpy>=1.26
matplotlib>=3.8
pydantic-settings>=2.0


streamlit==1.28.0
requests==2.31.0
pandas>=2.0
plotly>=5.11.0
scikit-learn>=1.2
